Rename StackFrame.CoreRT.cs to StackFrame.cs and move commons to shared
