+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -213,7 +213,6 @@
     <Compile Include="System\Diagnostics\DebuggerGuidedStepThroughAttribute.cs" />
     <Compile Include="System\Diagnostics\DebugAnnotations.cs" />
     <Compile Include="System\Diagnostics\StackFrame.cs" />
-    <Compile Include="System\Diagnostics\StackFrame.CoreRT.cs" />
     <Compile Include="System\Diagnostics\StackFrameExtensions.cs" />
     <Compile Include="System\Diagnostics\StackTrace.cs" />
     <Compile Include="System\Diagnostics\Tracing\EventSource_CoreRT.cs" />
